Extract name recommendation logic into a seperate type
